add winner and set auction status

+++ app/Http/Controllers/CustomerBidController.php
@@ -15,7 +15,8 @@ class CustomerBidController extends Controller
 {
     private $bidService;
     private $auctionService;
-    public function __construct(BidService $bidService, AuctionService $auctionService){
+    public function __construct(BidService $bidService, AuctionService $auctionService)
+    {
         $this->bidService = $bidService;
         $this->auctionService = $auctionService;
     }
@@ -25,46 +26,54 @@ class CustomerBidController extends Controller
         return view("bids.index", compact("bids"));
     }
     // for auction
-    public function show(Auction $auction){
-        if($this->bidService->checkBidsAvailabilityTime($auction))
-        {
-            $bids = $this->bidService->getAuctionBids($auction)->filter()->paginate(10);
-            $customers = $this->auctionService->getAuctionCustomers($auction)->paginate(10);
-        return view("bids.auction_bids", compact("bids", "auction","customers"));
-    }
-    else{
-        return redirect()->back()->with("error","you cannot enter the auction");
-    }
+    public function show(Auction $auction)
+    {
+        $bids = $this->bidService->getAuctionBids($auction)->filter()->paginate(10);
+        $customers = $this->auctionService->getAuctionCustomers($auction)->paginate(10);
+        if ($auction->status == 1) {
+            return view("bids.auction_bids", compact("bids", "auction", "customers"));
+        } else {
+            return view("bids.auction_bids", compact("bids", "auction", "customers"))->with("error", "auction was closed");
+        }
 
     }
 
-    public function store(BidRequest $request,Auction $auction)
+    public function store(BidRequest $request, Auction $auction)
     {
         $this->authorize("create", CustomerBid::class);
-        if($this->bidService->checkBidsAvailabilityTime($auction)){
-        $validated = $request->validated();
-        $validated['customer_id'] = Auth::user()->id;
-        $this->bidService->create($validated);
-        return redirect()->route("bids.show", $auction)->with("success","your bid was added");
-        }
-        else{
-        return redirect()->route("bids.show", $auction)->with("error","auction was closed");
+        if ($this->bidService->checkBidsAvailabilityTime($auction)) {
+            $validated = $request->validated();
+            $validated['customer_id'] = Auth::user()->id;
+            $this->bidService->create($validated);
+            return redirect()->route("bids.show", $auction)->with("success", "your bid was added");
+        } else {
+            return redirect()->route("bids.show", $auction)->with("error", "auction was closed");
         }
     }
 
     public function destroyAll(User $customer, Auction $auction)
     {
-        $bids = $this->bidService->getCustomerBidsForAuction($customer, $auction)->get();
-        foreach($bids as $bid){
-            $this->bidService->delete($bid);
+        if ($auction->status == 1) {
+            $bids = $this->bidService->getCustomerBidsForAuction($customer, $auction)->get();
+            foreach ($bids as $bid) {
+                $this->bidService->delete($bid);
+            }
+            return redirect()->route("customer.auction.bids", compact("bids", "auction"))->with("success", "you leaved the auction");
+        } else {
+            return redirect()->back()->with("error", "auction closed");
+
         }
-        return redirect()->route("bids.show", $auction)->with("success","you leaved the auction");
     }
     public function destroylatest(User $customer, Auction $auction)
     {
-        $bids = $this->bidService->getCustomerBidsForAuction($customer, $auction)->get();
-        // dd($bids[0]);
-        $this->bidService->delete($bids[0]);
-        return redirect()->back()->with("success","deleted latest bid successfully");
+        if ($auction->status == 1) {
+            $bids = $this->bidService->getCustomerBidsForAuction($customer, $auction)->get();
+            // dd($bids[0]);
+            $this->bidService->delete($bids[0]);
+            return redirect()->back()->with("success", "deleted latest bid successfully");
+        } else {
+            return redirect()->back()->with("error", "auction closed");
+
+        }
     }
 }
